backend/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.22
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
itsdangerous>=2.2.0
pytest>=8.0.0
httpx>=0.27.0
ruff>=0.8.0
